requirements.txt
absl-py==1.1.0
appnope==0.1.3
asttokens==2.0.5
backcall==0.2.0
backports.functools-lru-cache==1.6.4
debugpy==1.6.0
decorator==5.1.1
entrypoints==0.4
executing==0.8.3
flatbuffers==2.0
ipykernel==6.15.0
ipython==8.4.0
jax==0.3.13
jaxlib==0.3.10
jedi==0.18.1
jupyter-client==7.3.4
jupyter-core==4.10.0
matplotlib-inline==0.1.3
nest-asyncio==1.5.5
numpy==1.22.4
opt-einsum==3.3.0
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pip==22.1.2
prompt-toolkit==3.0.29
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.12.0
pyparsing==3.0.9
python-dateutil==2.8.2
pyzmq==23.1.0
scipy==1.8.1
setuptools==62.4.0
six==1.16.0
stack-data==0.3.0
tornado==6.1
traitlets==5.2.2.post1
typing_extensions==4.2.0
wcwidth==0.2.5
